requirements.txt
streamlit==1.27.1
tensorflow==2.13.0
numpy==1.26.0
pillow==10.0.0
matplotlib==3.8.1
pandas==2.1.0
seaborn==0.12.2
